+++ src/tensors/dispatch.h
@@ -152,6 +152,30 @@
       cpu::Function(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9); \
   }
 
+#define DISPATCH10(                                                             \
+  Function, Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10)        \
+namespace gpu {                                                                 \
+void Function(Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10);     \
+}                                                                               \
+namespace cpu {                                                                 \
+void Function(Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10);     \
+}                                                                               \
+static inline void Function(Arg1 arg1,                                          \
+                            Arg2 arg2,                                          \
+                            Arg3 arg3,                                          \
+                            Arg4 arg4,                                          \
+                            Arg5 arg5,                                          \
+                            Arg6 arg6,                                          \
+                            Arg7 arg7,                                          \
+                            Arg8 arg8,                                          \
+                            Arg9 arg9,                                          \
+                            Arg10 arg10) {                                      \
+  if(arg1->getBackend()->getDeviceId().type == DeviceType::gpu)                 \
+    gpu::Function(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10); \
+  else                                                                          \
+    cpu::Function(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10); \
+}
+
 #else
 
 #define DISPATCH1(Function, Arg1) \
@@ -246,6 +270,24 @@
                               Arg8 arg8,                                 \
                               Arg9 arg9) {                               \
     cpu::Function(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9); \
+  }
+
+#define DISPATCH10(                                                             \
+    Function, Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10)      \
+  namespace cpu {                                                               \
+  void Function(Arg1, Arg2, Arg3, Arg4, Arg5, Arg6, Arg7, Arg8, Arg9, Arg10);   \
+  }                                                                             \
+  static inline void Function(Arg1 arg1,                                        \
+                              Arg2 arg2,                                        \
+                              Arg3 arg3,                                        \
+                              Arg4 arg4,                                        \
+                              Arg5 arg5,                                        \
+                              Arg6 arg6,                                        \
+                              Arg7 arg7,                                        \
+                              Arg8 arg8,                                        \
+                              Arg9 arg9,                                        \
+                              Arg10 arg10) {                                    \
+    cpu::Function(arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10); \
   }
 
 #endif