+++ src/tensors/gpu/prod.h
@@ -6,6 +6,21 @@
 namespace marian {
 namespace gpu {
 
+void BiasAdd(marian::Tensor C,
+             const marian::Tensor& bias,
+             bool do_relu = false);
+
+void Affine(marian::Tensor C,
+            Ptr<Allocator> allocator,
+            const marian::Tensor& A,
+            const marian::Tensor& B,
+            const marian::Tensor& bias,
+            bool transA,
+            bool transB,
+            float beta = 0,
+            float scalar = 1,
+            bool do_relu = false);
+
 void Prod(marian::Tensor C,
           const marian::Tensor& A,
           const marian::Tensor& B,